Add a fallback to AMART to handle very big design spaces.
AMART Sampler must predict the distribution on all the design space at
each step. When the design space is very large this becomes problematic.
This commit adds an option where the distribution is only predicted in
a subset of points. For the moment the subset is choosen with an LHS.